fix(trips): address CI failure and PR review comments
- Fix CI: move formatDateRange and countdownLabel into
  features/trips/utils/tripDateUtils.ts with single-object params,
  satisfying the no-owned-max-params lint rule
- Refactor: move TripRow, EmptyState, and screen logic into
  features/trips/screens/UpcomingTripsScreen.tsx; slim route file
  to a thin feature-flag-gated entrypoint per feature-module convention
- i18n: replace hardcoded 'Today', 'Tomorrow', 'In N days', 'Plan a Trip',
  and empty-state copy with t() calls; add corresponding keys to en.json
